requirements.txt
apache-airflow==2.1.0
apache-airflow-providers-http==1.1.1
requests==2.25.1